Replace buildSrc with MossyPlugin
