+++ build.gradle
@@ -1,13 +1,13 @@
-import net.lopymine.mossy.MossyPlugin
+import net.lopymine.utils.MossyUtils
 
 plugins {
-	id("net.lopymine.mossy-plugin")
+	id "net.lopymine.mossy-plugin-core" version "$mossy_plugin_version"
 }
 
 mossyDependencies {
 	def mcVersion = stonecutter.current.version.toString()
 	minecraft = mcVersion
-	mappings = prop("build.yarn")
+	mappings = "mojang"
 	lombok = prop("base.lombok_version")
 	fabricApi = prop("build.fabric_api")
 	fabricLoader = prop("build.fabric_loader")
@@ -66,8 +66,6 @@ repositories {
 dependencies {
 	def mcVersion = stonecutter.current.version.toString()
 
-	mappings loom.officialMojangMappings()
-
 	// Fabric permission api
 	def fpaVersion = prop("build.fabric-permissions-api")
 	include(modImplementation('me.lucko:fabric-permissions-api:' + fpaVersion)){
@@ -76,7 +74,7 @@ dependencies {
 
 	// Cloth Config
 	def clothConfigVersion = prop("dep.cloth-config")
-	modApi("me.shedaniel.cloth:cloth-config-fabric:${stonecutter.compare("1.19", mcVersion) == 1 ? clothConfigVersion : MossyPlugin.substringBefore(clothConfigVersion, "+")}") {
+	modApi("me.shedaniel.cloth:cloth-config-fabric:${stonecutter.compare("1.19", mcVersion) == 1 ? clothConfigVersion : MossyUtils.substringBefore(clothConfigVersion, "+")}") {
 		exclude(group: "net.fabricmc.fabric-api")
 	}
 
@@ -122,6 +120,8 @@ stonecutter {
 	var buildTime = (int) (System.currentTimeMillis() / 10000) - 174611089
 	var buildCodeTime = String.format("%07x", buildTime & 0xFFFFFFF)
 	swap("build_code_time", "\"${buildCodeTime}\";")
+
+	dependency("yacl", MossyUtils.substringBefore(prop("dep.yacl"), "+"))
 }
