--- buildSettings/build.gradle
@@ -1,52 +1,0 @@
-plugins {
-	id 'java'
-	id 'java-gradle-plugin'
-}
-
-def mainModuleId = property("main_module_id").toString()
-def pluginMavenPackage = property("plugin_maven_package").toString()
-def pluginPackage = property("plugin_package").toString()
-def pluginClass = property("plugin_class").toString()
-def pluginVersion = property("plugin_version").toString()
-
-group = pluginPackage
-version = pluginVersion
-
-repositories {
-	mavenLocal()
-	mavenCentral()
-	gradlePluginPortal()
-	maven {
-		name "Fabric"
-		url "https://maven.fabricmc.net/"
-	}
-	maven {
-		name "Kiku Snapshots"
-		url "https://maven.kikugie.dev/snapshots"
-	}
-	maven {
-		name "Kiku Releases"
-		url "https://maven.kikugie.dev/releases"
-	}
-}
-
-dependencies {
-	compileOnly "org.projectlombok:lombok:${property("lombok_version")}"
-	annotationProcessor "org.projectlombok:lombok:${property("lombok_version")}"
-
-	implementation "com.google.code.gson:gson:${property("gson_version")}"
-
-	implementation "dev.kikugie.stonecutter:dev.kikugie.stonecutter.gradle.plugin:${property("stonecutter_version")}"
-}
-
-test {
-    useJUnitPlatform()
-}
-gradlePlugin {
-	plugins {
-		create(mainModuleId) {
-			id                  = "${pluginMavenPackage}.${mainModuleId}"
-			implementationClass = "${pluginPackage}.${pluginClass}"
-		}
-	}
-}
