--- buildSettings/gradle.properties
@@ -1,13 +1,0 @@
-# Gradle Plugin properties
-
-# Plugin Properties
-main_module_id = mossy-settings-plugin
-plugin_maven_package = net.lopymine
-plugin_package = net.lopymine.mossy.settings
-plugin_class = MossySettingsPlugin
-plugin_version = 1.0.0
-
-# Dependencies
-lombok_version = 1.18.42
-gson_version = 2.13.1
-stonecutter_version = 0.7.1
