--- buildSrc/gradle.properties
@@ -1,17 +1,0 @@
-# Gradle Plugin properties
-
-# Plugin Properties
-main_module_id = mossy-plugin
-stonecutter_module_id = mossy-plugin-stonecutter
-plugin_maven_package = net.lopymine
-plugin_package = net.lopymine.mossy
-plugin_class = MossyPlugin
-
-plugin_version = 1.0.0
-
-# Dependencies
-lombok_version = 1.18.42
-mod_publish_plugin_version = 0.7.4
-stonecutter_version = 0.7.1
-fabric_loom_version = 1.11-SNAPSHOT
-fletching_table_version = 0.1.0-alpha.17
