--- buildSrc/src/main/java/net/lopymine/mossy/extension/MossyProcessResourcesExtension.java
@@ -1,16 +1,0 @@
-package net.lopymine.mossy.extension;
-
-import lombok.Getter;
-import org.gradle.api.tasks.Input;
-
-import java.util.List;
-import org.jetbrains.annotations.Nullable;
-
-@Getter
-public class MossyProcessResourcesExtension {
-
-	@Input
-	@Nullable
-	List<String> expandFiles;
-
-}
