+++ gradle.properties
@@ -2,11 +2,15 @@
 #    Gradle Properties    #
 # # # # # # # # # # # # # #
 
-current_mossy_build_version = 1.0.0
-
 # Done to increase the memory available to gradle.
 org.gradle.jvmargs=-Xmx16G
 org.gradle.parallel=true
+
+# Mossy Plugin
+mossy_plugin_version = 1.1.1.0
+
+# Stonecutter
+dev.kikugie.stonecutter.hard_mode=true
 
 # Mod Properties
 data.mod_id = patpat
@@ -25,7 +29,7 @@ data.client_config_version = 1.0.1
 base.lombok_version = 1.18.42
 
 # Multi-Versions Properties
-multi_versions = 1.16.5 1.17 1.17.1 1.18 1.18.1 1.18.2 1.19 1.19.1 1.19.2 1.19.3 1.19.4 1.20 1.20.1 1.20.2 1.20.3 1.20.4 1.20.5 1.20.6 1.21 1.21.1 1.21.2 1.21.3 1.21.4 1.21.5 1.21.6 1.21.7 1.21.8 1.21.9 1.21.10
+multi_versions = 1.21.10 1.21.11
 
 # Build Dependencies
 build.yarn = [VERSIONED]
@@ -49,7 +53,7 @@ dep.proxlib = [VERSIONED]
 # Mod Loaders
 loaders = fabric
 # Versions Specifications
-versions_specifications = 1.16.5 1.17 1.17.1 1.18 1.18.1 1.18.2 1.19 1.19.1 1.19.2 1.19.3 1.19.4 1.20 1.20.1 1.20.2 1.20.3 1.20.4 1.20.5 1.20.6 1.21 1.21.1 1.21.2 1.21.3 1.21.4 1.21.5 1.21.6 1.21.7 1.21.8 1.21.9 1.21.10
+versions_specifications = 1.16.5 1.17 1.17.1 1.18 1.18.1 1.18.2 1.19 1.19.1 1.19.2 1.19.3 1.19.4 1.20 1.20.1 1.20.2 1.20.3 1.20.4 1.20.5 1.20.6 1.21 1.21.1 1.21.2 1.21.3 1.21.4 1.21.5 1.21.6 1.21.7 1.21.8 1.21.9 1.21.10 1.21.11
 # Version Type [RELEASE, BETA, ALPHA]
 version_type = RELEASE
